Moved lbPostsList in a separate file
